Fix nametag transform. nametag-transform -> billboard
